Covid_Project/pyproject.toml
[project]
name = "covid-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "pillow>=12.2.0",
    "streamlit>=1.56.0",
    "tensorflow>=2.21.0",
]
